+++ mobile/lib/screens/quiz/quiz_view.dart
@@ -1,9 +1,16 @@
 import 'package:app/models/models.dart';
+import 'package:app/screens/quiz/quiz_final_page.dart';
+import 'package:app/screens/quiz/quiz_title_page.dart';
 import 'package:app/screens/quiz/quiz_widgets.dart';
+import 'package:app/services/native_api.dart';
+import 'package:app/themes/app_theme.dart';
+import 'package:app/themes/colors.dart';
 import 'package:app/widgets/buttons.dart';
 import 'package:auto_size_text/auto_size_text.dart';
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
@@ -267,6 +274,104 @@ class _QuizQuestionsWidgetState extends State<QuizQuestionsWidget> {
           quiz: widget.quiz,
         );
       },
+    );
+  }
+}
+class QuizCard extends StatelessWidget {
+  const QuizCard(this.quiz, {super.key});
+  final Quiz quiz;
+
+  @override
+  Widget build(BuildContext context) {
+    return OutlinedButton(
+      style: OutlinedButton.styleFrom(
+        minimumSize: const Size.fromHeight(112),
+        foregroundColor: CustomColors.appColorBlue,
+        elevation: 0,
+        side: const BorderSide(
+          color: Colors.transparent,
+          width: 0,
+        ),
+        shape: const RoundedRectangleBorder(
+          borderRadius: BorderRadius.all(
+            Radius.circular(16),
+          ),
+        ),
+        backgroundColor: Colors.white,
+        padding: const EdgeInsets.fromLTRB(16.0, 8.0, 8.0, 8.0),
+      ),
+      onPressed: () async {
+        QuizQuestion? question = context.read<CurrentQuizQuestionCubit>().state;
+        if (question != null) {
+          dynamic response = await bottomSheetQuizQuestion(quiz, context);
+          if (response != null && response == true) {
+            //response = await QuizCompletionSheetContent();
+            response = await bottomSheetQuizConffeti(quiz, context);
+          }
+        } else {
+          context
+              .read<CurrentQuizQuestionCubit>()
+              .setQuestion(quiz.questions.first);
+          dynamic response = await bottomSheetQuizTitle(quiz, context);
+          if (response != null && response == true) {
+            //response = await QuizCompletionSheetContent();
+            response = await bottomSheetQuizQuestion(quiz, context);
+            if (response != null && response == true) {
+              //response = await QuizCompletionSheetContent();
+              response = await bottomSheetQuizConffeti(quiz, context);
+            }
+          }
+        }
+      },
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          SizedBox(
+            width: MediaQuery.of(context).size.width * 0.5,
+            height: 104,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Padding(
+                  padding: const EdgeInsets.only(bottom: 2),
+                  child: AutoSizeText(
+                    quiz.title,
+                    maxLines: 2,
+                    overflow: TextOverflow.ellipsis,
+                    style: CustomTextStyle.headline10(context),
+                  ),
+                ),
+                const Spacer(),
+                const QuizMessageChip(),
+              ],
+            ),
+          ),
+          const Spacer(),
+          SizedBox(
+            width: MediaQuery.of(context).size.width * 0.05,
+          ),
+          Container(
+            width: MediaQuery.of(context).size.width * 0.27,
+            height: 112,
+            decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(8.0),
+              image: DecorationImage(
+                fit: BoxFit.cover,
+                image: CachedNetworkImageProvider(
+                  quiz.imageUrl,
+                  cacheKey: quiz.imageUrlCacheKey(),
+                  cacheManager: CacheManager(
+                    CacheService.cacheConfig(
+                      quiz.imageUrlCacheKey(),
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
     );
   }
 }
