+++ mobile/lib/screens/quiz/quiz_widgets.dart
@@ -1,21 +1,12 @@
 import 'dart:math';
-
 import 'package:animated_text_kit/animated_text_kit.dart';
 import 'package:app/models/models.dart';
-import 'package:app/screens/quiz/quiz_final_page.dart';
-import 'package:app/screens/quiz/quiz_title_page.dart';
-import 'package:app/screens/quiz/quiz_view.dart';
-import 'package:app/services/native_api.dart';
 import 'package:app/themes/theme.dart';
 import 'package:auto_size_text/auto_size_text.dart';
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:confetti/confetti.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_svg/svg.dart';
-
 import '../../widgets/buttons.dart';
 
 class QuizSkipButton extends StatelessWidget {
@@ -50,105 +41,6 @@ class QuizSkipButton extends StatelessWidget {
           ),
           const SizedBox(
             width: 6,
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class QuizCard extends StatelessWidget {
-  const QuizCard(this.quiz, {super.key});
-  final Quiz quiz;
-
-  @override
-  Widget build(BuildContext context) {
-    return OutlinedButton(
-      style: OutlinedButton.styleFrom(
-        minimumSize: const Size.fromHeight(112),
-        foregroundColor: CustomColors.appColorBlue,
-        elevation: 0,
-        side: const BorderSide(
-          color: Colors.transparent,
-          width: 0,
-        ),
-        shape: const RoundedRectangleBorder(
-          borderRadius: BorderRadius.all(
-            Radius.circular(16),
-          ),
-        ),
-        backgroundColor: Colors.white,
-        padding: const EdgeInsets.fromLTRB(16.0, 8.0, 8.0, 8.0),
-      ),
-      onPressed: () async {
-        QuizQuestion? question = context.read<CurrentQuizQuestionCubit>().state;
-        if (question != null) {
-          dynamic response = await bottomSheetQuizQuestion(quiz, context);
-          if (response != null && response == true) {
-            //response = await QuizCompletionSheetContent();
-            response = await bottomSheetQuizConffeti(quiz, context);
-          }
-        } else {
-          context
-              .read<CurrentQuizQuestionCubit>()
-              .setQuestion(quiz.questions.first);
-          dynamic response = await bottomSheetQuizTitle(quiz, context);
-          if (response != null && response == true) {
-            //response = await QuizCompletionSheetContent();
-            response = await bottomSheetQuizQuestion(quiz, context);
-            if (response != null && response == true) {
-              //response = await QuizCompletionSheetContent();
-              response = await bottomSheetQuizConffeti(quiz, context);
-            }
-          }
-        }
-      },
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          SizedBox(
-            width: MediaQuery.of(context).size.width * 0.5,
-            height: 104,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.only(bottom: 2),
-                  child: AutoSizeText(
-                    quiz.title,
-                    maxLines: 2,
-                    overflow: TextOverflow.ellipsis,
-                    style: CustomTextStyle.headline10(context),
-                  ),
-                ),
-                const Spacer(),
-                const QuizMessageChip(),
-              ],
-            ),
-          ),
-          const Spacer(),
-          SizedBox(
-            width: MediaQuery.of(context).size.width * 0.05,
-          ),
-          Container(
-            width: MediaQuery.of(context).size.width * 0.27,
-            height: 112,
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(8.0),
-              image: DecorationImage(
-                fit: BoxFit.cover,
-                image: CachedNetworkImageProvider(
-                  quiz.imageUrl,
-                  cacheKey: quiz.imageUrlCacheKey(),
-                  cacheManager: CacheManager(
-                    CacheService.cacheConfig(
-                      quiz.imageUrlCacheKey(),
-                    ),
-                  ),
-                ),
-              ),
-            ),
           ),
         ],
       ),
@@ -212,7 +104,8 @@ class QuizDraggingHandle extends StatelessWidget {
 }
 
 class QuizProgressBar extends StatelessWidget {
-  const QuizProgressBar({super.key});
+  const QuizProgressBar(this.quiz, {super.key});
+  final Quiz quiz;
 
   @override
   Widget build(BuildContext context) {
@@ -223,7 +116,7 @@ class QuizProgressBar extends StatelessWidget {
         borderRadius: const BorderRadius.all(Radius.circular(10)),
         child: LinearProgressIndicator(
           color: CustomColors.appColorBlue,
-          value: 0.5,
+          value: quiz.activeQuestion / quiz.questions.length,
           backgroundColor: CustomColors.appColorBlue.withOpacity(0.24),
           valueColor: AlwaysStoppedAnimation<Color>(CustomColors.appColorBlue),
         ),
